requirements.txt
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
numpy==1.26.4
pandas==2.2.1
plotly==5.19.0
pyarrow==15.0.1
pymongo==4.6.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
requests==2.31.0
streamlit==1.32.0
